# Project-Mannager-System/docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: PM Daily Work Management API
  version: 1.0.0
  description: |
    API cho ứng dụng quản lý công việc hằng ngày của quản lý dự án phần mềm.

    ## Quy ước chung
    - Base path: `/api/v1`
    - Xác thực: `Authorization: Bearer <accessToken>` (JWT HS256, hết hạn 15 phút). Refresh token (7 ngày) chỉ gửi ở `/auth/refresh` và `/auth/logout`.
    - Thời gian: ISO-8601 UTC (`timestamptz`). Ngày (date): `yyyy-MM-dd`.
    - Phân trang: `page` (≥ 0, mặc định 0), `size` (1–100, mặc định 20), `sort` (whitelist theo từng endpoint, VD `fullName,asc`).
    - Lỗi: format thống nhất (BR-GEN-05) — `timestamp, status, error, code, message, path, fieldErrors, traceId`. Catalog mã lỗi: `docs/design/05-error-handling-design.md`.
    - Dữ liệu đã xóa mềm: không xuất hiện ở danh sách; truy cập trực tiếp trả 404.
    - Cập nhật bắt buộc kèm `version`; version cũ → 409.
    - Chi tiết từng module: `docs/api/01..12-*.md`.
servers:
  - url: http://localhost:8080/api/v1
    description: Local development
...
paths:
...
  /tasks:
    get:
      tags: [Task]
      summary: Danh sách công việc (nhiều filter)
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/Page' }
        - { $ref: '#/components/parameters/Size' }
        - { $ref: '#/components/parameters/Sort' }
        - { $ref: '#/components/parameters/ProjectIdQuery' }
        - { name: keyword, in: query, schema: { type: string } }
        - { name: assigneeId, in: query, schema: { type: string, format: uuid } }
        - { name: status, in: query, schema: { type: array, items: { $ref: '#/components/schemas/TaskStatus' } } }
        - { name: priority, in: query, schema: { type: array, items: { $ref: '#/components/schemas/TaskPriority' } } }
        - { name: type, in: query, schema: { type: array, items: { $ref: '#/components/schemas/TaskType' } } }
        - { name: tagId, in: query, schema: { type: string, format: uuid } }
        - { name: startDateFrom, in: query, schema: { type: string, format: date } }
        - { name: startDateTo, in: query, schema: { type: string, format: date } }
        - { name: dueDateFrom, in: query, schema: { type: string, format: date } }
        - { name: dueDateTo, in: query, schema: { type: string, format: date } }
        - { name: overdue, in: query, schema: { type: boolean } }
        - { name: blocked, in: query, schema: { type: boolean } }
      responses:
        '200':
          description: Trang kết quả
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TaskSummaryPage' }
        '400': { $ref: '#/components/responses/Validation400' }
        '403': { $ref: '#/components/responses/Forbidden403' }
    post:
      tags: [Task]
      summary: Tạo công việc (mã tự sinh PRJXXX-TASK-000001)
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TaskCreateRequest' }
      responses:
        '201':
          description: Tạo thành công
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TaskResponse' }
        '400': { $ref: '#/components/responses/Validation400' }
        '403': { $ref: '#/components/responses/Forbidden403' }
        '404': { $ref: '#/components/responses/NotFound404' }
  /tasks/my-tasks:
    get:
      tags: [Task]
      summary: Công việc được giao cho tôi (assignee = me)
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/Page' }
        - { $ref: '#/components/parameters/Size' }
        - { $ref: '#/components/parameters/Sort' }
        - { $ref: '#/components/parameters/ProjectIdQuery' }
      responses:
        '200':
          description: Trang kết quả
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TaskSummaryPage' }
  /tasks/today:
    get:
      tags: [Task]
      summary: Công việc hạn hôm nay (theo timezone user)
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/Page' }
        - { $ref: '#/components/parameters/Size' }
        - { $ref: '#/components/parameters/Sort' }
        - { $ref: '#/components/parameters/ProjectIdQuery' }
      responses:
        '200':
          description: Trang kết quả
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TaskSummaryPage' }
  /tasks/overdue:
    get:
      tags: [Task]
      summary: Công việc quá hạn chưa đóng
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/Page' }
        - { $ref: '#/components/parameters/Size' }
        - { $ref: '#/components/parameters/Sort' }
        - { $ref: '#/components/parameters/ProjectIdQuery' }
      responses:
        '200':
          description: Trang kết quả
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TaskSummaryPage' }
  /tasks/export:
    get:
      tags: [Task]
      summary: Xuất Excel danh sách task theo filter hiện tại
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/ProjectIdQuery' }
        - { name: keyword, in: query, schema: { type: string } }
        - { name: status, in: query, schema: { type: array, items: { $ref: '#/components/schemas/TaskStatus' } } }
        - { name: overdue, in: query, schema: { type: boolean } }
      responses:
        '200':
          description: File Excel
          content:
            application/vnd.openxmlformats-officedocument.spreadsheetml.sheet:
              schema: { type: string, format: binary }
        '400':
          description: EXPORT_LIMIT_EXCEEDED (quá 10.000 dòng)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '403': { $ref: '#/components/responses/Forbidden403' }
  /tasks/{id}:
    get:
      tags: [Task]
      summary: Chi tiết công việc
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/TaskIdPath' }
      responses:
        '200':
          description: Chi tiết task
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TaskResponse' }
        '403': { $ref: '#/components/responses/Forbidden403' }
        '404': { $ref: '#/components/responses/NotFound404' }
    put:
      tags: [Task]
      summary: "Cập nhật công việc (MEMBER: chỉ task của mình, trường status/progress/notes)"
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/TaskIdPath' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TaskUpdateRequest' }
      responses:
        '200':
          description: Đã cập nhật
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TaskResponse' }
        '400': { $ref: '#/components/responses/Validation400' }
        '403': { $ref: '#/components/responses/Forbidden403' }
        '404': { $ref: '#/components/responses/NotFound404' }
        '409': { $ref: '#/components/responses/Conflict409' }
    delete:
      tags: [Task]
      summary: Xóa mềm công việc (chặn khi còn task con)
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/TaskIdPath' }
      responses:
        '204': { description: Thành công, không body }
        '400':
          description: BAD_REQUEST khi còn task con
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '403': { $ref: '#/components/responses/Forbidden403' }
        '404': { $ref: '#/components/responses/NotFound404' }
  /tasks/{id}/assignee:
    put:
      tags: [Task]
      summary: Giao việc / đổi người thực hiện (sinh TASK_ASSIGNED)
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/TaskIdPath' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TaskAssigneeRequest' }
      responses:
        '200':
          description: Đã giao việc
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TaskResponse' }
        '400': { $ref: '#/components/responses/Validation400' }
        '403': { $ref: '#/components/responses/Forbidden403' }
        '404': { $ref: '#/components/responses/NotFound404' }
  /tasks/{id}/status:
    put:
      tags: [Task]
      summary: Chuyển trạng thái theo state machine (DONE ⇒ progress 100 + actualCompletedAt)
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/TaskIdPath' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TaskStatusRequest' }
      responses:
        '200':
          description: Đã chuyển trạng thái
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TaskResponse' }
        '400':
          description: INVALID_STATUS_TRANSITION / BLOCKER_REASON_REQUIRED / PROGRESS_REQUIRED_FOR_DONE
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '403': { $ref: '#/components/responses/Forbidden403' }
        '404': { $ref: '#/components/responses/NotFound404' }
  /tasks/{id}/progress:
    put:
      tags: [Task]
      summary: Cập nhật tiến độ 0–100
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/TaskIdPath' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TaskProgressRequest' }
      responses:
        '200':
          description: Đã cập nhật
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TaskResponse' }
        '400': { $ref: '#/components/responses/Validation400' }
        '403': { $ref: '#/components/responses/Forbidden403' }
        '404': { $ref: '#/components/responses/NotFound404' }
  /tasks/{id}/blocker:
    put:
      tags: [Task]
      summary: Đánh dấu / gỡ blocker (reason bắt buộc khi blocked)
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/TaskIdPath' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TaskBlockerRequest' }
      responses:
        '200':
          description: Đã cập nhật
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TaskResponse' }
        '400':
          description: BLOCKER_REASON_REQUIRED
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '403': { $ref: '#/components/responses/Forbidden403' }
        '404': { $ref: '#/components/responses/NotFound404' }
  /tasks/{id}/children:
    get:
      tags: [Task]
      summary: Danh sách task con
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/TaskIdPath' }
      responses:
        '200':
          description: Task con
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/TaskSummaryResponse' }
        '403': { $ref: '#/components/responses/Forbidden403' }
        '404': { $ref: '#/components/responses/NotFound404' }
  /tasks/{id}/tags:
    put:
      tags: [Task]
      summary: Gán tag cho task (thay thế toàn bộ)
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/TaskIdPath' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TagIdsRequest' }
      responses:
        '200':
          description: Đã cập nhật
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TaskResponse' }
        '403': { $ref: '#/components/responses/Forbidden403' }
        '404': { $ref: '#/components/responses/NotFound404' }
  /tasks/{id}/collaborators:
    put:
      tags: [Task]
      summary: Đặt danh sách người phối hợp (thuộc project)
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/TaskIdPath' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UserIdsRequest' }
      responses:
        '200':
          description: Đã cập nhật
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TaskResponse' }
        '400': { $ref: '#/components/responses/Validation400' }
        '403': { $ref: '#/components/responses/Forbidden403' }
        '404': { $ref: '#/components/responses/NotFound404' }
  /tasks/{id}/watchers:
    put:
      tags: [Task]
      summary: Đặt danh sách người theo dõi
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/TaskIdPath' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UserIdsRequest' }
      responses:
        '200':
          description: Đã cập nhật
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TaskResponse' }
        '400': { $ref: '#/components/responses/Validation400' }
        '403': { $ref: '#/components/responses/Forbidden403' }
        '404': { $ref: '#/components/responses/NotFound404' }
  /tasks/{id}/comments:
    get:
      tags: [Task]
      summary: Danh sách bình luận
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/TaskIdPath' }
      responses:
        '200':
          description: Bình luận
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/CommentResponse' }
        '403': { $ref: '#/components/responses/Forbidden403' }
        '404': { $ref: '#/components/responses/NotFound404' }
    post:
      tags: [Task]
      summary: Thêm bình luận (sinh TASK_COMMENTED cho assignee/watchers)
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/TaskIdPath' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CommentRequest' }
      responses:
        '201':
          description: Đã thêm
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CommentResponse' }
        '400': { $ref: '#/components/responses/Validation400' }
        '403': { $ref: '#/components/responses/Forbidden403' }
        '404': { $ref: '#/components/responses/NotFound404' }
  /tasks/{id}/comments/{commentId}:
    put:
      tags: [Task]
      summary: Sửa bình luận (chỉ tác giả)
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/TaskIdPath' }
        - { name: commentId, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CommentRequest' }
      responses:
        '200':
          description: Đã sửa
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CommentResponse' }
        '400': { $ref: '#/components/responses/Validation400' }
        '403': { $ref: '#/components/responses/Forbidden403' }
        '404': { $ref: '#/components/responses/NotFound404' }
    delete:
      tags: [Task]
      summary: Xóa bình luận (chỉ tác giả)
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/TaskIdPath' }
        - { name: commentId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '204': { description: Thành công, không body }
        '403': { $ref: '#/components/responses/Forbidden403' }
        '404': { $ref: '#/components/responses/NotFound404' }
  /tasks/{id}/attachments:
    get:
      tags: [Task]
      summary: Danh sách file đính kèm
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/TaskIdPath' }
      responses:
        '200':
          description: File đính kèm
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/AttachmentResponse' }
        '403': { $ref: '#/components/responses/Forbidden403' }
        '404': { $ref: '#/components/responses/NotFound404' }
    post:
      tags: [Task]
      summary: Upload file đính kèm (≤ 10MB, whitelist mime)
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/TaskIdPath' }
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
              required: [file]
      responses:
        '201':
          description: Đã upload
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AttachmentResponse' }
        '400':
          description: Sai mime type
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '403': { $ref: '#/components/responses/Forbidden403' }
        '404': { $ref: '#/components/responses/NotFound404' }
        '413': { $ref: '#/components/responses/PayloadTooLarge413' }
  /tasks/{id}/attachments/{attachmentId}:
    delete:
      tags: [Task]
      summary: Xóa file đính kèm (xóa khỏi storage + bản ghi)
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/TaskIdPath' }
        - { name: attachmentId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '204': { description: Thành công, không body }
        '403': { $ref: '#/components/responses/Forbidden403' }
        '404': { $ref: '#/components/responses/NotFound404' }
  /tasks/{id}/history:
    get:
      tags: [Task]
      summary: Lịch sử thay đổi (đọc từ audit_logs)
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/TaskIdPath' }
      responses:
        '200':
          description: Lịch sử
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/TaskHistoryEntry' }
        '403': { $ref: '#/components/responses/Forbidden403' }
        '404': { $ref: '#/components/responses/NotFound404' }
  # ============================= MEETING =============================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 0, default: 0 }
      description: Số trang (từ 0)
    Size:
      name: size
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
      description: Số bản ghi mỗi trang
    Sort:
      name: sort
      in: query
      schema: { type: string }
      description: 'Sắp xếp theo whitelist của endpoint, VD `fullName,asc`'
...
    ProjectIdQuery:
      name: projectId
      in: query
      schema: { type: string, format: uuid }
      description: Lọc theo dự án
...
  responses:
    Validation400:
      description: Dữ liệu không hợp lệ
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
    Forbidden403:
      description: Không đủ quyền hoặc ngoài phạm vi dữ liệu
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
  schemas:
...
    TaskStatus: { type: string, enum: [TODO, IN_PROGRESS, BLOCKED, REVIEW, DONE, CANCELLED] }
    TaskPriority: { type: string, enum: [LOW, MEDIUM, HIGH, CRITICAL] }
    TaskType: { type: string, enum: [FEATURE, BUG, IMPROVEMENT, TASK, OTHER] }
...
    TaskSummaryPage:
      allOf:
        - { $ref: '#/components/schemas/PageResponse' }
        - type: object
          properties:
            content:
              type: array
              items: { $ref: '#/components/schemas/TaskSummaryResponse' }
    # ---------- Meeting ----------
...